Codigos/2025/programacion IV/Tareas/6 Transformacion de aplicacion CLI a Aplicacion web con flask/requirements.txt
flask
redis
python-dotenv
